+++ flake.nix
@@ -2,6 +2,7 @@
   inputs = {
     nixpkgs.url = "github:NixOS/nixpkgs/nixpkgs-unstable";
     flake-utils.url = "github:numtide/flake-utils";
+    #clj-nix.url = "github:jlesquembre/clj-nix";
   };
 
   outputs =
@@ -23,6 +24,7 @@
       {
         overlays = final: prev: {
           datomic-pro = pkgs.callPackage ./pkgs/datomic-pro.nix { };
+          datomic-pro-peer = pkgs.callPackage ./pkgs/datomic-pro-peer.nix { };
           datomic-pro-container = pkgs.callPackage ./pkgs/datomic-pro-container-image.nix {
             imageTag = final.datomic-pro.version;
           };
@@ -34,6 +36,7 @@
         packages = {
           default = self.packages.${system}.datomic-pro;
           datomic-pro = pkgs.datomic-pro;
+          datomic-pro-peer = pkgs.datomic-pro-peer;
           datomic-pro-container = pkgs.datomic-pro-container;
           datomic-pro-container-unstable = pkgs.datomic-pro-container-unstable;
           datomic-generate-properties = pkgs.datomic-generate-properties;